refactor: move shared layout configurations to layout.shared.tsx
