+++ src/app/docs/layout.tsx
@@ -2,7 +2,7 @@ import { LargeSearchToggle } from 'fumadocs-ui/components/layout/search-toggle'
 import { DocsLayout } from 'fumadocs-ui/layouts/docs'
 import { Sparkles } from 'lucide-react'
 import type { ReactNode } from 'react'
-import { baseOptions, linkItems } from '@/app/layout.config'
+import { baseOptions, linkItems } from '@/lib/layout.shared'
 import { AISearchTrigger } from '@/components/fumadocs/ai'
 import { buttonVariants } from '@/components/ui/button'
 import { cn } from '@/lib/cn'
@@ -11,9 +11,11 @@ import 'katex/dist/katex.min.css'
 import DocsBackground from '@/components/docs-background'
 
 export default function Layout({ children }: { children: ReactNode }) {
+  const base = baseOptions();
+
   return (
     <DocsLayout
-      {...baseOptions}
+      {...base}
       tree={source.pageTree}
       // just icon items
       links={linkItems.filter((item) => item.type === 'icon')}
@@ -40,7 +42,7 @@ export default function Layout({ children }: { children: ReactNode }) {
         },
       }}
       nav={{
-        ...baseOptions.nav,
+        ...base.nav,
         children: (
           <AISearchTrigger
             className={cn(
